API/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
opencv-python==4.11.0.86
easyocr==1.7.2
annotated-types==0.7.0
easyocr==1.7.2
einops==0.8.1
matplotlib==3.10.3
numpy==1.26.4
openai==1.63.2
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pandas==1.5.3
pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
ultralytics==8.3.141
ultralytics-thop==2.0.14
wheel==0.45.1
yarl==1.20.0